rc_2.0.0: seeding process enhanced and multiple seed users functionality added
