--- prisma/seed.ts
@@ -1,39 +1,0 @@
-import { PrismaClient } from '@prisma/client';
-import bcrypt from 'bcryptjs';
-
-const prisma = new PrismaClient();
-
-async function main() {
-  // Check if admin already exists
-  const existingAdmin = await prisma.user.findUnique({
-    where: {
-      email: 'admin@example.com',
-    },
-  });
-
-  if (!existingAdmin) {
-    const hashedPassword = await bcrypt.hash('admin123', 10);
-
-    const admin = await prisma.user.create({
-      data: {
-        email: 'admin@example.com',
-        name: 'Admin User',
-        password: hashedPassword,
-        role: 'ADMIN',
-      },
-    });
-
-    console.log('Admin user created:', admin);
-  } else {
-    console.log('Admin user already exists');
-  }
-}
-
-main()
-  .catch((e) => {
-    console.error(e);
-    process.exit(1);
-  })
-  .finally(async () => {
-    await prisma.$disconnect();
-  });
