+++ prisma/seeds/index.ts
@@ -1,0 +1,15 @@
+import prisma from "./prisma";
+import { seedUsers } from "./users";
+
+async function main() {
+  await seedUsers();
+}
+
+main()
+  .catch((e) => {
+    console.error(e);
+    process.exit(1);
+  })
+  .finally(async () => {
+    await prisma.$disconnect();
+  });
